requirements.txt
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
boto3==1.35.67
botocore==1.35.67
cffi==1.17.1
click==8.1.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.5
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
libcst==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rich==13.9.4
s3transfer==0.10.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
strawberry-graphql==0.251.0
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
